requirements.txt
Flask
tensorflow[and-cuda]
numpy
pillow
matplotlib
opencv-python
